Forecast/requirements.txt
pandas>=2.0
numpy>=1.24
statsmodels>=0.14
matplotlib>=3.7
openpyxl>=3.1
streamlit>=1.35
python-calamine>=0.1
gspread>=6.0
google-auth>=2.0